+++ src/crate_name.rs
@@ -1,7 +1,7 @@
 use crate::{CrateData, Args};
 use crate::demangle::{self, SymbolName};
 
-const UNKNOWN: &str = "[Unknown]";
+pub const UNKNOWN: &str = "[Unknown]";
 
 pub(crate) fn from_sym(d: &CrateData, args: &Args, sym: &SymbolName) -> (String, bool) {
     let (mut name, is_exact) = from_sym_impl(d, sym);